src/app/config/__init__.py
import os
from dotenv import find_dotenv, load_dotenv


load_dotenv(find_dotenv())


class Development(object):
    TESTING = False 
    FLASK_DEBUG = os.getenv("FLASK_DEBUG")
    SECRET_KEY = os.getenv("SECRET_KEY")
    MONGO_URI = os.getenv("MONGO_URI")

class Production(object):
    TESTING = False 
    FLASK_DEBUG = os.getenv("FLASK_DEBUG")
    SECRET_KEY = os.getenv("SECRET_KEY")
    MONGO_URI = os.getenv("MONGO_URI")


app_config = {
    "development": Development,
    "production": Production
}